feat(fd2): Add the FD2 Modules and supporting code/structures (#2)
**Pull Request Description**

* Adds directories for JavaScript libraries (`library`), and Vue
components (`components`).
* Adds the `bin` directory with scripts for managing FarmData2.
* Adds the `modules` directory with _placeholders_ for the Drupal
modules:
  * `farm_fd2` - the main FarmData2 module.
* `farm_fd2_examples` - a module for canonical FD2 development examples.
  * `farm_fd2_school` - a module for the FarmData2 School activities.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.