chore: updates `addEntrypoint.bash` to use development branch (#11)
**Pull Request Description**

The `addEntrypoint.bash` script was written to use the `main` branch as
the base for the new feature branch in which to implement the entry
point. This had to be updated to match the use of `development` as the
default branch.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.