chore: updates entrypoint template to not use CounterExample component. (#12)
**Pull Request Description**

The entrypoint template `bin/templates/entrypoints/App.vue` used the
`CounterExample` component, which was removed from the repo. This PR
removes the use of it from the template.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.