chore: adds bootstrap-Vue-next to the entrypoint templates (#16)
**Pull Request Description**

Adds the bootstrap-vue-next UI library to the templates in `bin` that
are used to create new entrypoints. All entrypoints created will be able
to use BV next components.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.