chore: adds tools for documentation (#22)
**Pull Request Description**

Add the tools `jsdoc2md`, `vue-docgen-cli`, and `markdown-preview` along
with configuration and a script to generate documentation of the JS
modules in `library` and the Vue components in `components`.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.