chore: adds js libraries for farmOS and UI utilites (#23)
**Pull Request Description**

Adds `farmosUtil` and `uiUtil` to `library`.
- `farmosUtil` provides functions for working with a farmOS server via
the `farmos.js` library.
- `uiUtil` collects some reusable functions for working with
Bootstrap-Vue-Next.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.