feat: adds the CropSelector Vue component (#24)
**Pull Request Description**

Adds a Vue component for selecting a crop from the list of crops that
are in the `plant_type` taxonomy in the farmOS server.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.