feat: adds script for creating new Vue components (#27)
**Pull Request Description**

Adds the `addComponent.bash` script for creating a new Vue component in
`components`. The new component has a basic `.vue` file and three
`*.unit.cy.js` files that correspond to the 3 types of tests outlined in
`components/README.md`

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.