feat: adds validity to `CropSelector` and `DateSelector` components (#31)
**Pull Request Description**

Adds validity checking and styling to the `CropSelector` and
`DateSelector` components.

Closes #28

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.