feat: adds submit and reset buttons component (#34)
**Pull Request Description**

Adds a component containing the Submit and Reset buttons that will
appear at the bottom of each of the data entry forms.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.