fix: improves idioms for styling valid and invliad data elements (#35)
**Pull Request Description**

Updates the behavior of components so that:
  - Valid inputs are always marked with a green check
- Invalid inputs are only marked with a red x when the entry point sets
the `showInvalidStyling` prop.
- The `showInvalidStyling` prop also controls the display of the
`<BFormInvalidFeedback> element.

Factors this behavior out into a library function in `uiUtil.js` so that
it is common to all modules.

Updates the templates for new components.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.