feat: adds SelectorBase component (#36)
**Pull Request Description**

Factors out much of the functionality of a dropdown menu component into
the `SelectorBase` component so that it does not have to be repeated in
others (e.g. `CropSelector`, `LocationSelector`).

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.