fix: updates component templates to use standard validty styling (#37)
**Pull Request Description**

Updates the calls to the `uiUtil.js` library to use the new
`validityStyling` function to determine when to display the validity
styling on components.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.