chore: updates project config and npm dependencies (#40)
**Pull Request Description**

Adds and updates project configuration files and adds some npm
dependencies in preparation for other coming changes.

Note: There was an early feature branch that was very long running as
the architecture evolved and was refined. There were a number of PRs
merged into development from that branch that were eventually reset to
clean up and remove early missteps. The commits for those PRs no longer
exist in `development` but the closed PRs (#22 through #37)can still be
seen if desired.

This PR and the next several were created from the long running feature
branch to break some of the content of that branch into more logical
chunks. The division is not perfectly chronological, thus some of the
changes in these PRs may seem very prescient (e.g. gitignoring things
that don't yet exist but are added by later PRs, etc).

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.