build: updates entrypoints to include entry_point_name.html files (#41)
**Pull Request Description**

Each entry point only `index.html` as the file being served. This PR
adds `entry_point_name.html` as a copy of `index.html` to the built
module and reconfigures the vite configuration so that each built entry
point contains both an `index.html` and an `entry_point_name.html`
(which are identical). This was done because vite uses the
`entry_point_name.html` file to figure out what is to be included in the
built entry point, but the dev environment serves the `index.html` file.
Having both was a convenient way to make this all work.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.