feat(library): adds farmosUtil library for working with farmOS.js (#42)
**Pull Request Description**

Adds the `farmosUtil` library to facilitate working with the `farmOS.js`
library. This library file provides function for things like:
- Getting a `farm` instance
- Getting `Map`s of crops, users, locations, etc...
- Augmenting logs and assets with things like parents and owners.

This library will grow as additional functionality is needed by new
components and end points that are added.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.