feat(dev): adds script and templates for adding new components (#43)
**Pull Request Description**

Creates the `bin/addComponent.bash` script and associated templates for
adding new Vue components to the `components` library.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.