feat(comps): adds the BaseSelector component (#44)
**Pull Request Description**

Adds the `SelectorBase` component to `components`. The `SelectorBase`
component provides a dropdown select box that serves as a base component
for building specific _selectors_ such as a `CropSelector` or a
`LocationSelector`.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.