feat(docs): adds script for generating the library and component docs (#45)
**Pull Request Description**

Adds the `bin/makeDocs.bash` script that uses `jsdoc2md` and
`vue-docgen` and some custom scripting to generate documentation for the
`components` and `libraries`. The docs are generated with the `npm run
docs:gen` command. The docs are placed into the `docs` folder when
generated. The docs are viewed with the `npm run docs:view` command,
which opens a browser window and renders the markdown.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.