feat(comps): adds the CropSelector component (#46)
**Pull Request Description**

Adds the CropSelector component to `components`. The CropSelector
provides a dropdown select element for choosing one of the crops that is
contained in the farmOS vocabulary for `taxonomy_term--plant_type`.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.