feat(comps): adds the LocationSelector component (#47)
**Pull Request Description**

Adds the LocationSelector component to `components`. The
LocationSelector provides a dropdown select for choosing one of the
locations from the farmOS vocabularies:
- `type: 'asset--land', land_type: 'bed',`
- `type: 'asset--land', land_type: 'field',`
- `type: 'asset--structure', structure_type: 'greenhouse',`

The types of locations that are appear in the dropdown is controlled by
the component's props.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.