feat(comps): adds the DateSelector component (#48)
**Pull Request Description**

Adds the DateSelector component to `components`. The DateSelector
provides a date input UI element for the user to select a date.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.