feat(comps): adds the CommentBox component (#49)
**Pull Request Description**

Adds the CommentBox component to `components`. The CommentBox component
provides the comments text area element at the bottom of most input
forms.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.