feat(comps): adds the SubmitResetButtons component (#50)
**Pull Request Description**

Adds the SubmitResetButtons component to `components`. The
SubmitResetButtons component provides the Submit and Reset buttons that
appear at the bottom of all forms.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.