feat(comps): adds the NumericInput component (#52)
**Pull Request Description**

Add the NumericInput component to `components`. The NumericInput
component is a UI element that allows the user to enter a numeric input
value with a specified number of decimal places. The value entered is
automatically formatted to the specified number of decimal places.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.