fix(comps): updates the addComponent.bash script and templates (#53)
**Pull Request Description**

Makes updates to the script and templates used to add a new component.
- Adds a `styling` test template
- Documents `styling` tests in `components/README.md` draft
- Updates `addComponent.bash` to work with `styling` tests template
- Updates contents of other `.comp.cy.js` template files also.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.