feat(comps): adds TraySizeSelector component (#55)
**Pull Request Description**

Adds the TraySizeSelector to `components`. The TraySizeSelector provides
a UI element for selecting the number of cells in a tray used for tray
seeding. The `tray_size` taxonomy was added to the module config so that
it is created when the FarmData2 module is installed. Functions for
reading the `tray_size` vocabulary terms were also added to the
`library/farmosUtil/farmosUtil.js` library.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.