feat(comps): adds TextDisplay component (#56)
**Pull Request Description**

Adds the TextDisplay component to `components`. The TextDisplay
component is used to display non-editable labeled text in the UI. For
example, a value computed from other values entered into a form (e.g.
Total Seeds in the tray seeding form).

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.