fix(comps): ensures display is reactive to  prop changes (#57)
**Pull Request Description**

The text displayed was not reactive to changes in the `text` prop.
Changed the displayed text to be given by a computed property instead of
a method return value to fix this issue.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.