fix: gets the printFarmosLogs utility running (#58)
**Pull Request Description**

Gets the `printFarmosLogs.js` utility script working.
 - moves it to `bin`
 - changes `type` to `module` in `package.json`
- adjusts Cypress configs to be `.cjs` to be compatible with `type:
"module"`
 - revises `printFarmosLogs.js` to use schema to get the log types.

Closes #20

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.