fix(libs): patches the farmosUtil library to work in node and in browser (#59)
**Pull Request Description**

This PR patches the `library/farmosUtil` library so that it will work
both in the browser and in node. The issue was working with
`localStorage` and `sessionStorage` which do not exist in node. So a
separate library needed to be used to simulate them when running in
node. That had created issues with the js module types that was resolved
by allowing a simulated `localStorage` object to be passed in. In the
process the Cypress config for the unit tests was also changed so that
is uses vite instead of webpack.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.