chore: make eslintrc a .cjs and clean up linting errors (#60)
**Pull Request Description**

Setting `type: "module"` in `package.json` (in #58 -
9d29bc8b9be5f15a22098a910a3fbf34ef0013e1) broke linting because
`eslintrc.js` is common JS. This makes `eslintrc` a `.cjs` file instead
of a `.js` file so that it will work again.

In addition, this PR cleans up a bunch of linting errors that came up
when running: `npm run lint`

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.