feat(libs): adds utility functions for unit and log-category taxonomy terms (#65)
**Pull Request Description**

Adds functions and tests for them to the `farmosUtil.js` library for
interacting with the taxonomy terms for units and for log types.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.