refactor(libs): caches fetched data (#69)
**Pull Request Description**

All of the functions that make API requests have been updated to cache
their results in both a global variable (as a JS object) and in the
`sessionStorage` as stringified JSON. This ensures that when multiple
components or library functions make requests for the same data, an
actual API request only occurs once.

Closes #68 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.