refactor(lib): removes need to pass  to farmosUtil functoins (#71)
**Pull Request Description**

This PR refactors the `farmosUtil.js` library so that the utility
functions implicitly call `getFarmOSInstance` to get the `farm` instance
that they need. Prior to this callers were required to provide the
`farm` instance`. Because the `farmosUtil.js` library caches the `farm`
object, there is little overhead in having each utility method get its
own `farm` object instead of requiring the calling code to construct it.
This simplifies component and entry point development.

Closes #70

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.