fix(libs): fixes a race condition in getFarmOSInstance (#74)
**Pull Request Description**

This PR:
- Uses the `run-exclusive` package to fix a race condition that existed
in the `getFarmOSInstance` method when multiple components were making
concurrent calls to functions that created a farmOS instance. The
solution is essentially a mutex (or Java `synchronized` method) that
allows only one thread of execution in the `getFarmOSInstance` at a
time. This ensures that we do actually get only one farmOS instance.
- Enhances the testing of the caching of API results that is done by the
various `get...` functions in the `farmosUtil.js` library.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.