feat(libs): adds function for creating asset--plant records (#75)
**Pull Request Description**

Adds a function and associated tests to `farmosUtil.js` that simplifies
the creation of `asset--plant` records.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.