feat: adds support for lib.js to module entry points (#79)
**Pull Request Description**

This PR adds support to allow module entry points to include a `lib.js`
file for JavaScript functions and a collection of `lib.*.unit.cy.js`
tests for those functions. in doing so it:
- Updates the `bin/createEntryPoint.bash` script and the templates that
it uses when creating an entry point.
- These updates include a sample `lib.js` file with a sample
`lib.sample.unit.cy.js` test.
- In addition a `--dev` flag was added to create the new feature branch
from the current branch rather than `development`. This makes it easier
to work on the scripts and `templates`.
- Updates the `bin/test.bash` script to work with unit tests in the
module entry points.
- unifies the cypress configuration for the `farm_fd2`,
`farm_fd2_school`, `farm_fd2_examples` modules so that sym links are
unnecessary and this configuration only appears in the `modules`
directory.

Closes #25
Closes #26
Closes #66 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.

Co-authored-by: Ty Chermsirivatana <chermsit@dickinson.edu>