fix(libs): enable use of farmOS.js when served from module by farmOS (#81)
**Pull Request Description**

The farmOS.js library was designed to be used from outside of farmOS
itself (e.g. in another app, or in Node). It's default OAuth
authentication is not suitable for use within a page served from farmOS
as it would require login credentials to be supplied. As the farmOS user
accessing the page has already logged in this should not be necessary.

This PR, creates an `auth` config for farmOS.js that allows it to be
used within farmOS without providing credentials a second time. It also
maintains the ability to use farmOS.js from outside of farmOS by
providing login credentials.

Closes #62

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.