test: runs unit tests as component tests (#83)
**Pull Request Description**

Changes all unit tests to run as component tests instead of as e2e
tests. This is a better way to run the unit tests because they do not
need to access actual application pages.

Also, touches up a unit test to match a change in the units in the
sample database.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.