chore: updates the farmos2 image in docker_compose.yml (#85)
**Pull Request Description**

This PR updates the `docker/docker_compose.yml` file to use a new
version of the `farmos2` image. The new version includes some database
configuration information that simplifies the installation process.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.