fix: corrects group for docker/db directory (#86)
**Pull Request Description**

The `bin/installDB.bash` script assigned the incorrect group to the
`docker/db` directory. This PR corrects that so that the `docker/db`
directory has the group `fd2grp`. This enables that directory to be
written by the `fd2dev` user within the `fd2dev` container.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.