chore: migrates to farmOS v3 and postgres v13 (#90)
**Pull Request Description**

This PR updates the `docker/docker-compose.yml` file to use new images
for the farmOS and postgres.
- The farmOS image updates to v3.x.
- The postgres image updates to v13 (required by farmOS v3.x)

Closes https://github.com/FarmData2/FarmData2/issues/64

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.