fix: creates new farmOS instance of url, client or user change (#91)
**Pull Request Description**

Modified the `farmosUtil.getFarmOSInstance()` method so that if the
`hostURL`, `client` or `user` change then a new farmOS instance is
created and the old one is disposed of. This allows code to log in as
different users to test features that behave differently for different
users.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.