refactor: clarifies implementation of getFarmOSIinstance w.r.t. password grant authentication (#95)
**Pull Request Description**

This PR:
- adds the `cy.getAll` command `commands.js` for all of the cypress
configurations.
- refactors the `getFarmOSInstance()` function to more clearly delineate
between running in farmOS and running in the dev environment or Node.
This made the code much more clear and easier to see that the password
authentication is not being used when running inside of farmOS.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.