fix(library): patches farmosUtil.js and cleans up documentation (#97)
**Pull Request Description**

Makes a small fix to `farmosUtil.js` so that it more precisely detects
if it is running within farmOS or from Node or a dev server.

Also, cleans up and makes the documentation more complete. Including
adding `@category` tags, types for return values, and examples as
appropriate.

Closes #73

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.