test: modifies cypress tests initialize db before each test run (#100)
**Pull Request Description**

This PR changes the way the database is setup for cypress tests. Any
test that uses the database will reset the database to a released
version in the `before` hook. To support and implement this the PR:
- Revises the `installDB.bash` script to include command line
parameters.
- Updates all cypress tests that use the database to:
  -  reset the DB in `before`
- remove any database cleanup done in the tests (e.g. deleting
logs/assets)
- Updates relevant `README` drafts

Closes #99 
Closes #88 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.