feat(comps): adds farmOS permission checks to components (#106)
**Pull Request Description**

This PR:
- Adds functions to `farmosUtil.js` that check if the current user has a
given permission (e.g. create `asset--land`)
- Uses these functions in the components to determine if particular UI
elements should be displayed (e.g. `+` button for adding a new field).

Closes #63

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.