chore: a collection of small updates (#109)
**Pull Request Description**

This PR combines together a few small updates:
  - cspell dictionary was updated.
  - added conventional commit info to `CONTRIBUTING.md` 
- improved `installDB.bash` to allow pre-releases and to limit number of
releases shown in prompt.
  - cleaned up `printFarmosLogs.js` script

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.