feat(lib): updates to the library functions and tests (#110)
**Pull Request Description**

This PR: 
- documents the way the sample database is reset when running cypress
tests.
 - clarifies the operation of the `farmosUtil.crops.unit.cy.js` tests
- simplifies and fixes the `inFarmOS` function used by
`getFarmOSInstance`
 - removes database cleanup from `farmosUtil.getFarmInstance.unit.cy.js`
- removes comment describing the order dependency in
`farmosUtil.getFarmInstance.unit.cy.js` because the tests no longer have
an order dependency.
- adds the `hideToast` function to `uiUtil.js` to programmatically hide
a toast.

Partially addresses #101

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.