fix(comp): clean up components and their tests (#111)
**Pull Request Description**

This PR cleans up all of the components and their tests. Tests are
reorganized and augmented to match the structures and responsibilities
of the different test files as described in the `components/README.md`.

Closes #101 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.