fix: migrates fd2_school and fd2_examples to drupal 10 (#112)
**Pull Request Description**

This PR:
- migrates the `fd2_school` and `fd2_examples` to drupal 10
- updates the required farm/drupal modules for each of these modules
- renames/modifies basic e2e tests to include `e2e` in the filename

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.