feat(fd2): add the Tray Seeing entrypoint (#113)
**Pull Request Description**

This PR adds the Tray Seeding entry point to the farm_fd2 module to
collect data about tray seedings and create the appropriate farmOS
assets and logs. The assets and logs created are documented in
`docs/DataModel.md`.

Closes #10

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.