fix(fd2): adds drupal permission checking for entry point routes (#115)
**Pull Request Description**

This PR:
- Hides the Tray Seeding menu item for users without permission to
create plant assets, standard quantities and seeding logs.
- Updates the `addEntrypoint.bash` script and its `routing.yml` template
to allow the specification of the permissions necessary to see/use the
entry point.

Closes #105

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.