ci: fixes drupal-releae hook so it copies all changed files (#116)
**Pull Request Description**

Patches the `drupal-release` git hook so that all changed files are
copied into the drupal farmdata2 module repo.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.