test(fd2): tests permission based access to tray seeding (#117)
**Pull Request Description**

Adds a test that checks that the guest user does not have access to the
Tray Seeding form when running on the live farmOS server. The test is
skipped when running on the dev or preview servers.

This PR also adds `@cypress/skip-test` as a dev dependency to allow
dynamic skipping of tests.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.