chore(lib): remove debugging output (#118)
**Pull Request Description**

Removes `console.log` statements from `farmosUtil.js` that were added
for debugging and were not removed before committing.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.