perf(fd2): Imporve user permissions checking (#120)
**Pull Request Description**

Implements an API endpoint (`api/permissions`) that returns an object
specifying if the current user has or does not have each of the
permissions that FarmData2 needs to know. The permissions that are
checked are specified in the
`modules/farm_fd2/src/module/Controller/FD2_Controller.php` file. The
permissions fetching and checking functions in the `farmosUtil.js`
library were changed to use this new API endpoint. The components that
use permissions to check if UI elements are to be displayed were
modified to use the new library functions.

Closes #119 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.