Refactor caching (#123)
**Pull Request Description**

This PR factors the caching functionality in `farmosUtil.js` out into
its own function and then refactors all of the functions that use that
caching functionality so that they use the factored out function. This
reduces the overall amount of code and test code. Additional tests were
added that test just the caching functionality and the corresponding
tests for each of the refactored functions was removed.

Closes #92

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.