test: patch the tests to work with smaller sample db (#128)
**Pull Request Description**

The sample db was trimmed down in size by eliminating many locations and
many crops. This PR adjusts all of the tests (comp, unit, e2e) to pass
with those changes.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.