feat(fd2): adds the seeding parent menu (#131)
**Pull Request Description**

This PR adds the "Seeding" parent menu to the "FarmData2" menu. It moves
the "Tray Seeding" input form to this parent menu and will provide a
parent for the additional seeding related features (direct seeding,
cover crop seeding, seeding reports, etc...)

Related to #129 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.