refactor: move asset, quantity, log creation into `farmosUtil.js` (#138)
**Pull Request Description**

This PR:
- Adds functions to the `farmosUtil.js` library that are able to create,
fetch and delete `asset--plant`, `quantity--standard` and `log--seeding`
- Adds tests for the functions added to `farmosUtil.js`
- Refactors `lib.js` in the Tray Seeding entry point to use these
generic functions.
- Refactors the unit tests for the Tray Seeding `lib.js`

Overall this greatly simplifies the creation of new entry points as they
will not have to repeat the code for creating these types of assets,
quantities and logs. As additional entry points are added the
`farmosUtil.js` library will need to expand to include functions for any
new types of assets, quantities, logs, etc. that are needed.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.