feat(fd2): adds a direct seeding entry point (#132)
**Pull Request Description**

This PR adds an entry point for recording direct seeding events. It
includes:
- Creation of the entry point
- Creation of all of the new components that it requires
- Additions to the `farmosUtil.js` library that are required
  - Fetching equipment and creating maps
  - Creating, fetching, deleting:
- `asset--plant`, `quantity--standard`, `log-seeding`, `log-activity`
records.
- Adds "Direct Seeding" entry point to the FarmData2 -> Seeding menu.
- Update to the organization of end point tests described in
`modules/README.md`
- A collection of odds and ends that came up as the entry point was
being created

Closes #129 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.