fix(fd2): Added WSL2 Support (#121)
**Pull Request Description**

Closes #76 ```fd2-up.linux.bash``` does not setup proper groups

Added bash commands in /bin/fd2-up.linux.bash to add necessary groups
and grpid in WSL2 environment to start fd2
_Include a "Closes #" line for the issue your PR closes._

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.