docs: updates draft of `INSTALL.md` with some additinoal details (#143)
**Pull Request Description**

Updates the `INSTALL.md` and also modifies the `defaultCommandTimeout`
for Cypress so that tests will pass on some slower machines.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.