fix(fd2): allows the use of multiple log categories (#148)
**Pull Request Description**

This PR makes the use of log categories consistent with the Survey Stack
conventions for entering direct seeding data into farmOS but ensuring
that logs include their `log_category` as well as any FarmData2 specific
log categories.

Closes #146 
Partially addresses #108

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.