fix(fd2): creates log categories on module install (#149)
**Pull Request Description**

This PR adds the `farm_fd2.install` file to the `farm_fd2` module. This
file is executed by drupal when the module is installed and it creates
the log categories that are used by FarmData2 (if they do not already
exist).

Additional code can be added to `farm_fd2.install` to create additional
log categories or terms in other taxonomies (e.g. `units`).

Closes #139

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.