fix(fd2): migrates creation of unit taxonomy terms to install (#151)
**Pull Request Description**

Move the creation of the `unit` taxonomy terms used by FarmData2 from
the sample database to the `farm_fd2.install` file so that they are
created whenever the FarmData2 module is installed.

Related to #139 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.