test(lib): updates tests for log categories created by install (#153)
**Pull Request Description**

Patches the log category tests for the `farmosUtil` library to match the
log categories that are added by `farm_fd2.install` rather than the
previous ones added in the creation of the sample database.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.