test(fd2): rewrite tray seeding tests to follow new conventions (#155)
**Pull Request Description**

The conventions for testing end points evolved as the direct seeding
entry point was being written and tested. This PR goes back and revises
the tray seeding tests so that they now follow the convention that has
been established. See `modules/README.md` for information about the
testing conventions.

Closes #140
Closes #145

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.