fix(dev): fd2-up.linux.bash GID numbering logic (#162)
**Pull Request Description**

Fixed `fd2-up.linux.bash` to only read the number in the `fd2grp.gid`
file. The previous code was reading the comments as well, which led to
errors with `getent`.

Also added logic to handle the case when the GID 3000 is already used by
the system.

Install script now confirmed to be working on both Debian (my machine)
and Ubuntu (VirtualBox).

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.