feat(examples): adds a DateSelector example for FD2 Examples (#172)
**Pull Request Description**

Adds a `Components` section to the `FD2 Examples` menu.
Adds a `DateSelector` example to the `Components` menu.

Bumps fd2dev image to new version that includes Firefox add-ons for
viewing example source code.

Closes #171 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.