fix(examples): adds control for date prop to DateSelector example (#173)
**Pull Request Description**

Adds control for changing the `date` `prop` to the `DateSelector`
example.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.