fix(dev): moves database file to ~/.fd2/ so it persists in docker volume (#174)
**Pull Request Description**

Moves the location to which sample database asset files are downloaded
from `/var/tmp` to `~/.fd2/` so that the downloaded file is persisted in
the Docker volume for the user's home directory. This ensures that the
`installDB.bash -c` script will be able to install the most recent
database after the dev environment is taken down and brought back up.

This PR also:
- Updates the `fd2-up.bash` and `fd2-up-linux.bash` scripts to work with
absolute paths so that they can be run from any directory within the
repo.
- Updates `fd2-up.bash` so that it does not change to `docker` directory
until necessary.

Closes #122

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.