feat(examples: adds example for the `PickerBase` component (#178)
**Pull Request Description**

Closes #177 

Also: 
- Fixes a bug in `PickerBase` ensuring that if an option is removed it
is also removed from the picked items as well.
- Adds some css for formatting tables of `props` and event payloads in
the examples.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.