feat(examples): adds a BedPicker example (#176)
**Pull Request Description**

Closes #175 

Also:
- Fixes a bug in BedPicker where going to a location with no beds did
not clear picked beds. Fixed by using `v-show` instead of `v-if` for the
`PickerBase` and adding a `<div>` around `PickerBase` `<template>`
content.
- Adds test for `BedPicker` visibility when there are no beds.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.