test(comp): fixes update event test in LocationSelector (#179)
**Pull Request Description**

Fixes the `LocationSelector` events test for the case when the location
changes. This was broken by changes to the `PickerBase` and `BedPicker`
components.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.