fix(examples): tweaks BedPicker example (#181)
**Pull Request Description**

Updates the `BedPicker` example so that:
- The second table heading matches the others.
- The `ALF` radio button is initially selected to reflect the initial
state of the component.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.