fix(examples): links the BedPicker example in jsDoc and examples page (#182)
**Pull Request Description**

- Adds a link to the Components example page for the `BedPicker`
- Updates the `BedPicker` jsDoc to
  - Link to the live example and source code
  - revise the static example code in the jsDoc

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.