Fix validity styling (#183)
**Pull Request Description**

Updates the computation of `showValidityStyling` in components such that
invalid styling is not shown when non-required values are empty.

Closes #165 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.