feat(comp): add picklist base component (#187)
**Pull Request Description**

Add a new component that presents a list of items in tabular form and
allows the user to select one or more of the rows for later operation.
This component will be useful for selecting plantings to be transplanted
or harvested (among other things).

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.