refactor: switches to pgrep for finding servers (#194)
Switches the `killZombieServers.bash` script to use `pgrep` instead of
`grep`ing `ps -ax` as it gives a cleaner result.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.