feat(fd2): submit tray and direct seedings forms as a transaction (#212)
**Pull Request Description**

When submitting a form (e.g. tray seeding) multiple quantities, logs and
assets are created. This can result in an issue of there is an error
when creating one or more of them. When such an error occurs, ideally
all logs, assets or quantities that were created should be deleted. The
`runTransaction` function added to `farmosUtil.js` facilitates this by
making it easy to specify a set of operations with `create` and `delete`
functions. It then runs each operation's `create` function in sequence.
if one fails, it will run the `delete` function for each operation that
had previously succeeded.

This is not a perfect solution however, because if the error is caused
by a network outage, then the delete operations probably will not
succeed either. But it is a step in the right direction.

The implementations of `submitForm` in the `lib.js` for the Tray Seeding
and Direct Seeding entry points has been refactored to use
`runTransaction` instead of managing the undo manually.

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.