refactor(fd2): Generalize transaction (#214)
**Pull Request Description**

Modifies the `runTransaction` function to use the more general function
names `do` and `undo` for the operations. Also, refactors the Tray
Seeding and Direct Seeding `lib.js` files to use the new `do` and `undo`
function names for transaction operations.

Closes #213 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.