feat(fd2): Add transplanting entry point (#200)
**Pull Request Description**

This PR:
- Adds a Transplanting endpoint to the FarmData2 menu that provides data
entry for recording a transplanting event.
- Includes enhancements to the `PicklistBase` component as implemented
during the creation of the Transplanting endpoint.
- Adds `TransplantingPicklist` component.
- Adds functions to the `farmosUtiil.js` library related to
transplanting.
- Other miscellaneous changes as implemented during the creation of the
Transplanting endpoint.

Closes #198 
Closes #196 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.