fix(fd2): updates sticky fields in direct seeding and transplanting forms (#228)
**Pull Request Description**

The Beds, Bed feet and Area fields in the Direct Seeding and
Transplanting forms were changed so that they are not sticky. The Area
field required an update to the `NumericInput` component to allow it to
be blank. Finally an extraneous duplicate `page-loaded` element was
removed from the Transplanting form.

Closes #224 
Closes #226 

---

**Licensing Certification**

FarmData2 is a [Free Cultural
Work](https://freedomdefined.org/Definition) and all accepted
contributions are licensed as described in the LICENSE.md file. This
requires that the contributor holds the rights to do so. By submitting
this pull request **I certify that I satisfy the terms of the [Developer
Certificate of Origin](https://developercertificate.org/)** for its
contents.